% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_participant.R
\name{choose_participant}
\alias{choose_participant}
\title{Select participant by id}
\usage{
choose_participant(id)
}
\arguments{
\item{id}{Participant identifier.}
}
\description{
Select participant dataset (list) from participant list.
}
