% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_df.R
\name{format_df}
\alias{format_df}
\title{Format numeric values for printing data frames}
\usage{
format_df(df, digits = 2, tryInt = TRUE, align_right = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{digits}{to how many decimals should numeric data be rounded?}

\item{tryInt}{check if numeric data are integer like and do not round if so}

\item{align_right}{total number of characters to be used for numeric data, use this for nice alignment, NOT IMPLEMENTED}
}
\value{
Formatted data.frame.
}
\description{
Format numeric values by converting them to character.
}
