% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_cb.R
\name{mini_cb}
\alias{mini_cb}
\title{Minimal codebook from data.frame}
\usage{
mini_cb(df, auto.type = TRUE, content.type = FALSE)
}
\arguments{
\item{df}{data.frame.}

\item{auto.type}{logical. Should the column's data type be filled in
automatically?}

\item{content.type}{logical. Should there be an extra row to describe what
this codebook contains?}
}
\value{
Returns a data.frame with one row per column of the original
data.frame.
}
\description{
Create a minimal codebook df from an input df. For each column
of the input df, one row will be created. Save the codebook df as a text
file in the format of your choice (e.g. csv or json) and fill it out.
There will be columns for the variable/ measurement name, the data type, the
unit of measurement, a description and an annotation.
}
\examples{
# example code
df <- data.frame(id = 1:4,
                 condition = factor(LETTERS[1:4]),
                 score = c(2.1, 4.0, NA, 3.3))
cb <- mini_cb(df, auto.type = TRUE, content.type = TRUE)
cb
# write(jsonlite::toJSON(cb, pretty = TRUE), "cb_path.json")

}
