% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_p_lst.R
\docType{data}
\name{p_lst}
\alias{p_lst}
\title{List of participants}
\format{
An object of class \code{list} of length 2.
}
\usage{
p_lst
}
\description{
An example list dataset to demonstrate and test functions.
}
\keyword{datasets}
