% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_trtctrl_df.R
\docType{data}
\name{trtctrl_df}
\alias{trtctrl_df}
\title{Treatment - Control Dataframe}
\format{
A data frame with 5 variables (one row per observation unit):
\describe{
\item{\code{id}}{Participant identifier, factor.}
\item{\code{grp}}{Experimental group;
control ("ctrl") or treatment ("trt"), factor.}
\item{\code{age_yrs}}{Age in years, numeric.}
\item{\code{gender_fmx}}{Gender; female ("f"), male ("m"), non-binary ("x").}
\item{\code{score_pre}}{Pretest score on scale XYZ, numeric.}
\item{\code{score_post}}{Posttest score on scale XYZ, numeric.}
}
}
\usage{
trtctrl_df
}
\description{
An example dataset to demonstrate and test functions.
}
\keyword{datasets}
