% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputoutput_tsv.R
\name{write_tsv}
\alias{write_tsv}
\title{Tab-stop separated data output}
\usage{
write_tsv(x, ...)
}
\arguments{
\item{x}{object to be written, matrix or data frame.}

\item{...}{argument to \code{write.table}: \code{sep} cannot be altered.}
}
\value{
Write tsv file.
}
\description{
Wrapper for write.table() to write tab-stop separated text file (tsv).
}
